Change EN locale when fluid mining is not yet researched
Factorio defaults to a message that says "Uranium mining not researched" when trying to place a drill on any ore that requires fluids. This is obviously not correct for Py Hard Mode.

This updates the message to reference "Usage of fluid in mining drills", which is the unlock shown in the "Steam power" tech unlocked by producing iron plates.